requirements.txt
easyocr
numpy
pillow
groq
ipython
google-generativeai